requirements.txt
functions-framework
python-dotenv
flask
twilio
openai
google-cloud-secret-manager
google-cloud-storage

